Problem with json format

+++ src/Controller/ControllerApi.php
@@ -15,6 +15,7 @@ class ControllerApi extends Controller
     /**
      * @Route("/note/api/get", name="noteApiGet")
      * @Method({"GET"})
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|Response
      */
     public function index()
     {
@@ -25,7 +26,7 @@ class ControllerApi extends Controller
         $response->headers->set('Content-Type', 'application/json');
         return $response;
     }
-    /**TODO : cette fonction ne marche pas.
+    /** TODO : cette fonction ne marche pas.
       * @Route("/note/api/post", name="noteApiPost")
       * @Method({"POST"})
       * @param Request $request
@@ -51,11 +52,19 @@ class ControllerApi extends Controller
 
     /**
      * @Route("/note/api/delete/{id}", name="noteApiDelete")
-     * @Method({"DELETE", "GET"})
+     * @Method({"DELETE"})
      * @param $id
      */
     public function deleteNote($id){
       $note = $this->getDoctrine()->getRepository(Note::class)->find($id);
+      if(!$note){
+        return new JsonResponse(
+                  array(
+                      'status' => 'UNKNOWN',
+                      'message' => "Note doesn't exist"
+                  )
+              );
+      }
       $entityManager=$this->getDoctrine()->getManager();
       $entityManager->remove($note);
       $entityManager->flush();
@@ -64,6 +73,30 @@ class ControllerApi extends Controller
                     'status' => 'DELETED',
                     'message' => 'This note has been deleted'
                 )
-            );;
+            );
+    }
+    /**
+     * @Route("/note/api/put/{id}", name="noteApiPut")
+     * @Method("PUT")
+     * @param $id
+     */
+     public function editNote($id, Request $request){
+      $entityManager=$this->getDoctrine()->getManager();
+      $note = $this->getDoctrine()->getRepository(Note::class)->find($id);
+      $content = $request->getContent();
+      if(empty($content))
+      {
+        return new JsonResponse(array('status'=>'EMPTY','message'=>'The body of this request is empty.'));
+      }
+      if (!$note) {
+        return new JsonResponse(array('status'=>'UNKNOWN', 'message'=>"Note doesn't exist"));
+      }
+      $contentRequest = $this->get('jms_serializer')->deserialize($content, Note::class, 'json');
+      $note->setTitle($contentRequest->getTitle());
+      $note->setContent($contentRequest->getContent());
+      $note->setDate($contentRequest->getDate());
+      $note->setCategory($contentRequest->getCategory());
+      $entityManager->flush();
+      return new JsonResponse(array('status'=>'MODIFIED', 'message'=>'Note successfully modified'));
     }
 }
